requirements.txt
Flask>=2.3
SQLAlchemy>=2.0
mysql-connector-python>=8.3
python-dotenv>=1.0
